refactor snowpack/src/commands/dev.ts into smaller files/classes
