+++ jest.config.js
@@ -5,6 +5,7 @@ module.exports = {
     '<rootDir>/test/create-snowpack-app/test-install', // don’t run tests inside our mock create-snowpack-app install
     '<rootDir>/www', // docs has its own tests
   ],
+  testRegex: '(/__tests__/.*|(\\.|/)(test|spec))\\.js$', // `tsx?` and `jsx` files are not transformed, thus only test `js` file
   setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
   testEnvironment: 'node',
 };
