--- snowpack/.npmignore
@@ -1,2 +1,0 @@
-src
-
